requirements.txt
tensorflow==2.15.0
streamlit==1.29.0
pillow==10.1.0
numpy==1.26.2
